+++ webargs/flaskparser.py
@@ -19,14 +19,11 @@ Example: ::
     def index(args):
         return 'Hello ' + args['name']
 """
-import logging
-
 import flask
 from werkzeug.exceptions import HTTPException
 
 from webargs import core
 
-logger = logging.getLogger(__name__)
 
 def abort(http_status_code, **kwargs):
     """Raise a HTTPException for the given http_status_code. Attach any keyword
@@ -82,7 +79,6 @@ class FlaskParser(core.Parser):
         """Handles errors during parsing. Aborts the current HTTP request and
         responds with a 422 error.
         """
-        logger.error(error)
         status_code = getattr(error, 'status_code', self.DEFAULT_VALIDATION_STATUS)
         abort(status_code, messages=error.messages, exc=error)