+++ tests/test_aiohttp/test_aiohttpparser.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 from webargs import fields
+from marshmallow import Schema
 
 import asyncio
 
@@ -85,6 +86,25 @@ def test_parsing_cookies(app, client):
     assert res.status_code == 200
     assert res.json == {'mycookie': 'foo'}
 
+def test_parse_with_callable(app, client):
+
+    class MySchema(Schema):
+        foo = fields.Field(missing=42)
+
+    def make_schema(req):
+        return MySchema(context={'request': req})
+
+    @asyncio.coroutine
+    def echo_parse(request):
+        args = yield from parser.parse(make_schema, request)
+        return jsonify(args)
+
+    app.router.add_route('GET', '/factory', echo_parse)
+
+    res = client.get('/factory')
+    assert res.status_code == 200
+    assert res.json == {'foo': 42}
+
 
 def test_use_args(app, client):
 
@@ -102,6 +122,26 @@ def test_use_args(app, client):
     res = client.get('/?name=Joe')
     assert res.status_code == 200
     assert res.json == {'name': 'Joe'}
+
+
+def test_use_args_with_callable(app, client):
+
+    class MySchema(Schema):
+        foo = fields.Field(missing=42)
+
+    def make_schema(req):
+        return MySchema(context={'request': req})
+
+    @asyncio.coroutine
+    @use_args(make_schema)
+    def echo_use_args(request, args):
+        return jsonify(args)
+
+    app.router.add_route('GET', '/use_args', echo_use_args)
+
+    res = client.get('/use_args')
+    assert res.status_code == 200
+    assert res.json == {'foo': 42}
 
 
 def test_use_kwargs_on_method(app, client):